backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
channels==4.0.0

# Database
psycopg2-binary==2.9.9

# Firebase
firebase-admin==6.2.0

# Payments
stripe==7.6.0

# WebSocket
channels-redis==4.2.0
daphne==4.0.0

# Documentation
drf-yasg==1.21.7
uritemplate==4.1.1

# Utils
python-dotenv==1.0.1
Pillow==10.1.0
requests==2.31.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.1 